requirements.txt
pytest~=8.3.5
playwright~=1.51.0
pytest-playwright~=0.3.0
allure-pytest~=2.14.0
pytest-xdist~=3.5.0
requests~=2.31.0
python-dotenv~=1.0.0
psycopg2-binary~=2.9.9
clickhouse-driver~=0.2.6
boto3~=1.34.71
botocore~=1.34.71
